Struct state_tree::StateTree

pub struct StateTree<'db, S> { /* fields omitted */ }

State tree implementation using hamt. This structure is not threadsafe and should only be used in sync contexts.

Implementations

impl<'db, S> StateTree<'db, S> where
S: BlockStore,

pub fn new(
store: &'db S,
version: StateTreeVersion
) -> Result<Self, Box<dyn StdError>>

pub fn new_from_root(store: &'db S, c: &Cid) -> Result<Self, Box<dyn StdError>>

Constructor for a hamt state tree given an IPLD store

pub async fn import_state_tree<R: AsyncRead + Send + Unpin>(
store: &'db S,
reader: R
) -> Result<Cid, String>

Imports a StateTree given an AsyncRead that has data in CAR format

pub fn store(&self) -> &S

Retrieve store reference to modify db.

pub fn get_actor(
&self,
addr: &Address
) -> Result<Option<ActorState>, Box<dyn StdError>>

Get actor state from an address. Will be resolved to ID address.

pub fn set_actor(
&mut self,
addr: &Address,
actor: ActorState
) -> Result<(), Box<dyn StdError>>

Set actor state for an address. Will set state at ID address.

pub fn lookup_id(
&self,
addr: &Address
) -> Result<Option<Address>, Box<dyn StdError>>

Get an ID address from any Address

pub fn delete_actor(&mut self, addr: &Address) -> Result<(), Box<dyn StdError>>

Delete actor for an address. Will resolve to ID address to delete.

pub fn mutate_actor<F>(
&mut self,
addr: &Address,
mutate: F
) -> Result<(), Box<dyn StdError>> where
F: FnOnce(&mut ActorState) -> Result<(), String>,

Mutate and set actor state for an Address.

pub fn register_new_address(
&mut self,
addr: &Address
) -> Result<Address, Box<dyn StdError>>

Register a new address through the init actor.

pub fn snapshot(&mut self) -> Result<(), String>

Add snapshot layer to stack.

pub fn clear_snapshot(&mut self) -> Result<(), String>

Merges last two snap shot layers.

pub fn revert_to_snapshot(&mut self) -> Result<(), String>

Revert state cache by removing last snapshot

pub fn flush(&mut self) -> Result<Cid, Box<dyn StdError>>

Flush state tree and return Cid root.

pub fn for_each<F>(&self, f: F) -> Result<(), Box<dyn StdError>> where
F: FnMut(Address, &ActorState) -> Result<(), Box<dyn StdError>>,
S: BlockStore,