Trait state_tree::StateTree

pub trait StateTree {
    fn get_actor(&self, addr: &Address) -> Result<Option<ActorState>, String>;

    fn set_actor(
        &mut self, 
        addr: &Address, 
        actor: ActorState
    ) -> Result<(), String>;

    fn lookup_id(&self, addr: &Address) -> Result<Address, String>;

    fn delete_actor(&mut self, addr: &Address) -> Result<(), String>;

    fn mutate_actor<F>(
        &mut self, 
        addr: &Address, 
        mutate: F
    ) -> Result<(), String>
    where
        F: FnOnce(ActorState) -> Result<ActorState, String>;

    fn register_new_address(
        &mut self, 
        addr: &Address, 
        actor: ActorState
    ) -> Result<Address, String>;

    fn snapshot(&mut self) -> Result<Cid, String>;

    fn revert_to_snapshot(&mut self, cid: &Cid) -> Result<(), String>;
}

Interface to allow for the retreival and modification of Actors and their state

Required methods

fn get_actor(&self, addr: &Address) -> Result<Option<ActorState>, String>

Get actor state from an address. Will be resolved to ID address.

fn set_actor(&mut self, addr: &Address, actor: ActorState) -> Result<(), String>

Set actor state for an address. Will set state at ID address.

fn lookup_id(&self, addr: &Address) -> Result<Address, String>

Get an ID address from any Address

fn delete_actor(&mut self, addr: &Address) -> Result<(), String>

Delete actor for an address. Will resolve to ID address to delete.

fn mutate_actor<F>(&mut self, addr: &Address, mutate: F) -> Result<(), String> where
F: FnOnce(ActorState) -> Result<ActorState, String>,

Mutate and set actor state for an Address.

fn register_new_address(
&mut self,
addr: &Address,
actor: ActorState
) -> Result<Address, String>

Register a new address through the init actor.

fn snapshot(&mut self) -> Result<Cid, String>

Persist changes to store and return Cid to revert state to.

fn revert_to_snapshot(&mut self, cid: &Cid) -> Result<(), String>

Revert to Cid returned from snapshot

Loading content...